Add files via upload
Project Description: The Project is based on Instagram User Analytics; dataset is given with all variables of the project. Initially I will go through the given dataset and see the observation of each and every variable table and formulate the output for all perspective tasks.  

Approach: I will study the whole dataset for the given Instagram users, Analyse  the different data inside the table and note down the total number of observations for different data available and I will also execute the data for the required task which is mentioned for each task.

Tech-Stack used: My SQL Workbench -8.0 CE
I used My SQL Workbench, it's a software for SQL database playground for testing, debugging and executing the code. The user interface is easy to understand, and we can execute the code within minimal time. 

INSIGHTS: With the help of application My SQL Workbench, I have imported the file and executed it. Initially I looked in to all the variables of the database. (Users, tags,photos,photo_tags,likes,follows,comments). with respect to the target task, I have written code line to obtain all the data in tables for tasks.

I have executed the file into two software's Microsoft SQL server and 
Db fiddle, but those application have some constraints with the file, and it was easily done in My SQL workbench.